Removed db.
