+++ src/controllers/watch_time_controller.rs
@@ -6,8 +6,10 @@ use std::sync::{Mutex};
 use crate::config::dbconfig::establish_connection;
 use crate::DbPool;
 use crate::models::episode::Episode;
+use crate::models::podcast_history_item::PodcastHistoryItem;
 use crate::models::user::User;
 use crate::mutex::LockResultExt;
+use crate::service::mapping_service::MappingService;
 
 #[utoipa::path(
 context_path="/api/v1",
@@ -23,7 +25,8 @@ Responder {
 
 
     let podcast_episode_id = podcast_watch.0.podcast_episode_id.clone();
-    DB::log_watchtime(&mut conn.get().unwrap(),podcast_watch.0, requester.unwrap().username.clone())
+    PodcastHistoryItem::log_watchtime(&mut conn.get().unwrap(),podcast_watch.0, requester.unwrap().username
+        .clone())
         .expect("Error logging watchtime");
     log::debug!("Logged watchtime for episode: {}", podcast_episode_id);
     HttpResponse::Ok()
@@ -36,13 +39,15 @@ responses(
 tag="watchtime"
 )]
 #[get("/podcast/episode/lastwatched")]
-pub async fn get_last_watched(db: Data<Mutex<DB>>, conn: Data<DbPool>, requester: Option<web::ReqData<User>>) -> impl
+pub async fn get_last_watched(db: Data<Mutex<DB>>, conn: Data<DbPool>, requester:
+Option<web::ReqData<User>>, mapping_service:Data<Mutex<MappingService>>) -> impl
 Responder {
 
     let designated_username = requester.unwrap().username.clone();
     let mut db = db.lock().ignore_poison();
-    let last_watched = db.get_last_watched_podcasts(&mut establish_connection(), designated_username
-        .clone()).unwrap();
+    let last_watched = PodcastHistoryItem::get_last_watched_podcasts(&mut establish_connection(),
+                                                     designated_username
+        .clone(), mapping_service.lock().ignore_poison().clone()).unwrap();
 
     let episodes = Episode::get_last_watched_episodes(designated_username, &mut conn.get().unwrap(),
     );
@@ -80,6 +85,7 @@ tag="watchtime"
 #[get("/podcast/episode/{id}")]
 pub async fn get_watchtime(id: web::Path<String>, conn: Data<DbPool>, requester: Option<web::ReqData<User>>) -> impl Responder {
     let designated_username = requester.unwrap().username.clone();
-    let watchtime = DB::get_watchtime(&mut conn.get().unwrap(),&id, designated_username).unwrap();
+    let watchtime = PodcastHistoryItem::get_watchtime(&mut conn.get().unwrap(),&id, designated_username)
+        .unwrap();
     HttpResponse::Ok().json(watchtime)
 }
