+++ src/main.rs
@@ -58,7 +58,9 @@ mod service;
 use crate::db::DB;
 use crate::gpodder::parametrization::get_client_parametrization;
 use crate::gpodder::routes::get_gpodder_api;
+use crate::models::podcasts::Podcast;
 use crate::models::session::Session;
+use crate::models::settings::Setting;
 use crate::models::web_socket_message::Lobby;
 use crate::service::environment_service::EnvironmentService;
 use crate::service::file_service::FileService;
@@ -91,7 +93,7 @@ pub fn run_poll(
     mut podcast_service: PodcastService,
     mut podcast_episode_service: PodcastEpisodeService) {
     //check for new episodes
-    let podcats_result = DB::get_all_podcasts(&mut establish_connection()).unwrap();
+    let podcats_result = Podcast::get_all_podcasts(&mut establish_connection()).unwrap();
     for podcast in podcats_result {
         if podcast.active {
             let podcast_clone = podcast.clone();
@@ -173,7 +175,7 @@ async fn main() -> std::io::Result<()> {
         let polling_interval = env.get_polling_interval();
         scheduler.every(polling_interval.minutes()).run(|| {
             let conn = &mut establish_connection();
-            let settings = DB::new().unwrap().get_settings(conn);
+            let settings = Setting::get_settings(conn);
             match settings {
                 Some(settings) => {
                     if settings.auto_update {
@@ -196,7 +198,7 @@ async fn main() -> std::io::Result<()> {
             sessions");
             let mut db = DB::new().unwrap();
             let mut podcast_episode_service = PodcastEpisodeService::new();
-            let settings = db.get_settings(conn);
+            let settings = Setting::get_settings(conn);
             match settings {
                 Some(settings) => {
                     if settings.auto_cleanup {
@@ -371,14 +373,14 @@ pub fn get_secure_user_management() ->Scope{
 pub fn insert_default_settings_if_not_present() {
     let mut db = DB::new().unwrap();
     let conn = &mut establish_connection();
-    let settings = db.get_settings(conn);
+    let settings = Setting::get_settings(conn);
     match settings {
         Some(_) => {
             info!("Settings already present");
         }
         None => {
             info!("No settings found, inserting default settings");
-            db.insert_default_settings(conn);
+            Setting::insert_default_settings(conn);
         }
     }
 }
