+++ src/models/episode.rs
@@ -12,9 +12,10 @@ use diesel::query_builder::QueryBuilder;
 use crate::db::DB;
 use crate::{DbConnection, MyQueryBuilder};
 use crate::dbconfig::schema::podcast_episodes::dsl::podcast_episodes;
+use crate::models::models::PodcastWatchedEpisodeModelWithPodcastEpisode;
 use crate::models::podcast_episode::PodcastEpisode;
+use crate::models::podcast_history_item::PodcastHistoryItem;
 use crate::models::podcasts::Podcast;
-use crate::models::models::{PodcastHistoryItem, PodcastWatchedEpisodeModelWithPodcastEpisode};
 
 #[derive(Serialize, Deserialize, Debug,Queryable, QueryableByName,Insertable, Clone, ToSchema)]
 pub struct Episode{
