+++ src/models/settings.rs
@@ -1,7 +1,12 @@
 use crate::dbconfig::schema::*;
 use diesel::prelude::{AsChangeset, Identifiable, Insertable, Queryable};
+use diesel::{OptionalExtension, RunQueryDsl};
 use crate::service::environment_service::OidcConfig;
 use utoipa::ToSchema;
+use crate::DbConnection;
+use crate::utils::do_retry::do_retry;
+use diesel::insert_into;
+use crate::constants::constants::DEFAULT_SETTINGS;
 
 #[derive(
     Serialize, Deserialize, Queryable, Insertable, Debug, Clone, Identifiable, AsChangeset,ToSchema
@@ -30,4 +35,43 @@ pub struct ConfigModel {
     pub basic_auth: bool,
     pub oidc_configured: bool,
     pub oidc_config: Option<OidcConfig>
+}
+
+
+impl Setting{
+
+    pub fn get_settings(conn: &mut DbConnection) -> Option<Setting> {
+        use crate::dbconfig::schema::settings::dsl::*;
+
+        settings
+            .first::<Setting>(conn)
+            .optional()
+            .unwrap()
+    }
+
+    pub fn update_settings(setting: Setting, conn:&mut DbConnection) -> Setting {
+        use crate::dbconfig::schema::settings::dsl::*;
+        let setting_to_update = settings
+            .first::<Setting>(conn)
+            .expect("Error loading settings");
+        do_retry(||{diesel::update(&setting_to_update)
+            .set(setting.clone())
+            .get_result::<Setting>(conn)})
+            .expect("Error updating settings")
+    }
+
+    pub fn insert_default_settings(conn: &mut DbConnection) {
+        use crate::dbconfig::schema::settings::dsl::*;
+        use diesel::ExpressionMethods;
+        do_retry(||{insert_into(settings)
+            .values((
+                id.eq(1),
+                auto_download.eq(DEFAULT_SETTINGS.auto_download),
+                auto_cleanup.eq(DEFAULT_SETTINGS.auto_cleanup),
+                auto_cleanup_days.eq(DEFAULT_SETTINGS.auto_cleanup_days),
+                podcast_prefill.eq(DEFAULT_SETTINGS.podcast_prefill))
+            )
+            .execute(conn)})
+            .expect("Error setting default values");
+    }
 }
