+++ src/service/notification_service.rs
@@ -1,6 +1,6 @@
 use crate::db::DB;
 use crate::DbConnection;
-use crate::models::models::Notification;
+use crate::models::notification::Notification;
 
 #[derive(Clone)]
 pub struct NotificationService {
@@ -12,13 +12,13 @@ impl NotificationService {
         }
     }
 
-    pub fn get_unread_notifications(&mut self, mut db:DB, conn:&mut DbConnection)
+    pub fn get_unread_notifications(&mut self, conn:&mut DbConnection)
         ->Result<Vec<Notification>, String>{
-        db.get_unread_notifications(conn)
+        Notification::get_unread_notifications(conn)
     }
 
     pub fn update_status_of_notification(&mut self, id: i32, status: &str, mut db:DB,conn:&mut DbConnection) -> Result<(),
         String> {
-        db.update_status_of_notification(id, status, conn)
+        Notification::update_status_of_notification(id, status, conn)
     }
 }
